+++ client/lib/pages/trail_page.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_svg/svg.dart';
+import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:trailquest/pages/generate_trail_page.dart';
 import 'package:trailquest/my_trail_list.dart';
 import 'package:trailquest/widgets/trail_cards.dart';
@@ -18,15 +19,15 @@ class TrailPage extends StatefulWidget {
 
 class _TrailPageState extends State<TrailPage> {
   //final List<TrailCard> myTrails = MyTrailList;
-  //final List<TrailCard> friendsTrails = FriendTrailList;
+  final List<TrailCard> friendsTrails = FriendTrailList;
 
   List<TrailCard> myTrails = [];
-  List<TrailCard> friendsTrails = []; 
+  //List<TrailCard> friendsTrails = [];
 
   @override
   void initState() {
     super.initState();
-    fetchTrails();
+    fetchUserTrailsFromServer();
   }
 
   //TODO: if we can get i fron the users saved trails in database
@@ -47,6 +48,73 @@ class _TrailPageState extends State<TrailPage> {
       friendsTrails.add(dummyTrail);
     }
     setState(() {});
+  }
+
+  //TODO: Fetch from database, simulated data below
+  void fetchUserTrailsFromServer() {
+    var trailsData = [
+      {
+        "trailName": "Test trail 1",
+        "totalDistance": 31.0,
+        "totalTime": 0.363849765258216,
+        "statusEnvironment": "Both",
+        "avoidStairs": false,
+        "hilliness": 0.09058952331542969,
+        "coordinates": [
+          {"latitude": 59.85697, "longitude": 17.62699},
+          {"latitude": 59.85695, "longitude": 17.62704},
+          {"latitude": 59.8569, "longitude": 17.62693},
+          {"latitude": 59.85692, "longitude": 17.62688},
+          {"latitude": 59.8569, "longitude": 17.62693},
+          {"latitude": 59.85686, "longitude": 17.62699},
+          {"latitude": 59.8569, "longitude": 17.62693},
+          {"latitude": 59.85695, "longitude": 17.62704},
+          {"latitude": 59.8569, "longitude": 17.62712},
+          {"latitude": 59.85695, "longitude": 17.62704}
+        ]
+      },
+      {
+        "trailName": "Test trail 2",
+        "totalDistance": 16.0,
+        "totalTime": 0.18779342723004697,
+        "statusEnvironment": "Both",
+        "avoidStairs": false,
+        "hilliness": 0.013283729553222656,
+        "coordinates": [
+          {"latitude": 59.86001, "longitude": 17.64235},
+          {"latitude": 59.85997, "longitude": 17.64227},
+          {"latitude": 59.86001, "longitude": 17.64235},
+          {"latitude": 59.86004, "longitude": 17.64242},
+          {"latitude": 59.86007, "longitude": 17.64247},
+          {"latitude": 59.86001, "longitude": 17.64235}
+        ]
+      },
+    ];
+
+    List<TrailCard> newTrails = [];
+    for (var data in trailsData) {
+      List<LatLng> trailCoordinates = [];
+      var coordinatesData = data['coordinates'];
+      if (coordinatesData != null && coordinatesData is List) {
+        trailCoordinates = List<LatLng>.from(coordinatesData
+            .map((coord) => LatLng(coord['latitude'], coord['longitude'])));
+      }
+
+      newTrails.add(TrailCard(
+          name: data['trailName'].toString(),
+          lengthDistance: data['totalDistance'] as double,
+          lengthTime: data['totalTime'] as double,
+          natureStatus: data['statusEnvironment'] as String,
+          stairs: data['avoidStairs'] as bool,
+          heightDifference: data['hilliness'] as double,
+          isSaved: true,
+          isCircular: false,
+          coordinates: trailCoordinates));
+    }
+
+    setState(() {
+      myTrails = newTrails;
+    });
   }
 
   @override
